modules: rename built-in plugins to "builtins"
Updated relevant plugin handling, documentation/examples, & test files.
