+++ sopel/builtins/admin.py
