+++ sopel/builtins/announce.py
