+++ sopel/builtins/calc.py
