+++ sopel/builtins/choose.py
