+++ sopel/builtins/clock.py
