+++ sopel/builtins/countdown.py
