+++ sopel/builtins/currency.py
