+++ sopel/builtins/dice.py
