+++ sopel/builtins/emoticons.py
