+++ sopel/builtins/find.py
