+++ sopel/builtins/help.py
