+++ sopel/builtins/isup.py
