+++ sopel/builtins/ping.py
