+++ sopel/builtins/rand.py
