+++ sopel/builtins/safety.py
