+++ sopel/builtins/search.py
