+++ sopel/builtins/seen.py
