+++ sopel/builtins/tell.py
