+++ sopel/builtins/translate.py
