+++ sopel/builtins/unicode_info.py
