+++ sopel/builtins/version.py
