+++ sopel/builtins/wiktionary.py
