+++ test/builtins/test_builtins_announce.py
@@ -1,7 +1,7 @@
 """Tests for Sopel's ``announce`` plugin"""
 from __future__ import annotations
 
-from sopel.modules import announce
+from sopel.builtins import announce
 
 
 def test_chunks():
