+++ test/vcr/builtins/currency/test_example_exchange_cmd_1.yaml
