+++ test/vcr/builtins/search/test_example_duck_0.yaml
