+++ test/vcr/builtins/translate/test_example_tr2_0.yaml
