+++ test/vcr/builtins/translate/test_example_tr2_2.yaml
